requirements.txt
pandas
scipy
matplotlib
pymongo
python-dotenv
python-telegram-bot
Pillow

